Update quantiles.csv
ISI Web of Knowledge's Esencial Science Indicators Percentile Table for Geosciences.

"The Percentiles table displays data on the minimum number of citations needed to meet six percentile breakdowns within the scientific fields.
The Percentiles table is divided into sections for each scientific field with the All Fields section combining all data. Within each section are rows for the six percentile breakdowns (0.01%, 0.10%, 1.00%, 10.00%, 20.00% and 50.00%). Data for the percentiles are given for the last 10 years with the All Years column using all data for those years."